+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ** Changelog for Dasbhoard
 
+v0.2
+- prototype to add react routes
+
 v0.1
 - scrapped SB dashboard react template, unusable with other react applications
 - began to build custom dashboard template from scratch
+- added base footer / header from MDB